Added menu option to load config.xml file
Added option to load config.xml file
PTZcommandButtons - if menu checkbox is on, the window will display even if there are no available presets to display. Display can be turned off using the menu checkbox. This is the same behavior as in PTZControl presets combobox, otherwise it causes other display issues which /i can't solve.